Time Reset
Time goes back to set time when export is finished. All input is blocked when exporting, and a progress bar and dialogue is shown. Need to fix the first frame being blank and change the name of the folder, as well as fixing setting times, as it doesnt properly set the time.